bot_one/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
python-decouple==3.8
python-telegram-bot==20.2
pytz==2023.3
rfc3986==1.5.0
sniffio==1.3.0
yarl==1.8.2
